requirements.txt
pbkdf2==1.3
PyNaCl==0.3.0
qrcode==5.1
Pillow==2.7.0
zbar==0.10
PyVimeo==0.3.0
